+++ packages/eslint-plugin/src/util/astUtils.ts
@@ -1,7 +1,7 @@
 import {
-  TSESTree,
-  AST_TOKEN_TYPES,
   AST_NODE_TYPES,
+  AST_TOKEN_TYPES,
+  TSESTree,
 } from '@typescript-eslint/experimental-utils';
 
 const LINEBREAK_MATCHER = /\r\n|[\r\n\u2028\u2029]/;
@@ -15,6 +15,17 @@ function isNotOptionalChainPunctuator(
   token: TSESTree.Token | TSESTree.Comment,
 ): boolean {
   return !isOptionalChainPunctuator(token);
+}
+
+function isNonNullAssertionPunctuator(
+  token: TSESTree.Token | TSESTree.Comment,
+): boolean {
+  return token.type === AST_TOKEN_TYPES.Punctuator && token.value === '!';
+}
+function isNotNonNullAssertionPunctuator(
+  token: TSESTree.Token | TSESTree.Comment,
+): boolean {
+  return !isNonNullAssertionPunctuator(token);
 }
 
 /**
@@ -31,9 +42,32 @@ function isOptionalOptionalChain(
   );
 }
 
+/**
+ * Returns true if and only if the node represents logical OR
+ */
+function isLogicalOrOperator(node: TSESTree.Node): boolean {
+  return (
+    node.type === AST_NODE_TYPES.LogicalExpression && node.operator === '||'
+  );
+}
+
+/**
+ * Determines whether two adjacent tokens are on the same line
+ */
+function isTokenOnSameLine(
+  left: TSESTree.Token,
+  right: TSESTree.Token,
+): boolean {
+  return left.loc.end.line === right.loc.start.line;
+}
+
 export {
-  LINEBREAK_MATCHER,
+  isNonNullAssertionPunctuator,
+  isNotNonNullAssertionPunctuator,
   isNotOptionalChainPunctuator,
   isOptionalChainPunctuator,
   isOptionalOptionalChain,
+  isTokenOnSameLine,
+  isLogicalOrOperator,
+  LINEBREAK_MATCHER,
 };